refactor: 💡 add error handling
